+++ lib/kafka_impl/util.ex
@@ -1,5 +1,6 @@
 defmodule KafkaImpl.Util do
   def extract_offset([%KafkaEx.Protocol.Offset.Response{partition_offsets: [%{offset: [offset]}]}]), do: {:ok, offset}
+  def extract_offset([%KafkaEx.Protocol.Offset.Response{partition_offsets: [%{offset: []}]}]), do: :no_offset
   def extract_offset(error), do: {:error, "Can't extract offset: #{inspect error}"}
 
   def extract_messages([%{partitions: partitions}]) do