+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 CHANGELOG inspiration from http://keepachangelog.com/.
 
 ## Unreleased
+* Add last_committed_offset_for
+* Move agent handling to Store
+* Move test helper functions to TestHelper
 
 ## [0.1.0] - October 25, 2016
 * Initial release